Backend/requirements.txt
Flask
Flask-Cors
python-docx
PyMuPDF
google-generativeai
gspread
oauth2client
python-dotenv
